Added DateValidationUtils.scala object and moved all date validation code to same file.
